型定義の統合・整理を実施。domain/event.tsにEvent, EventType, EventStatus, EVENT_TYPE, EVENT_STATUSを集約し、各所の重複を排除。全てのlint・テストを通過済み。モデル: GPT-4.1
